Add cookie message provider
Also update MVC sample to show how to use.
The provider does not work as-expected if an exception is thrown since no cookie is set in the response.
I created an issue (https://github.com/aspnet/Mvc/issues/4707) to see if this is a framework problem or if there is just a better way to handle this scenario.